log4cxx 0.11.0

+++ Libraries/3rd/log4cxx/source/src/main/cpp/bytearrayinputstream.cpp
@@ -29,27 +29,34 @@ using namespace std;
 IMPLEMENT_LOG4CXX_OBJECT(ByteArrayInputStream)
 
 ByteArrayInputStream::ByteArrayInputStream(const std::vector<unsigned char>& bytes) :
-    buf(bytes), pos(0) {
+	buf(bytes), pos(0)
+{
 }
 
 
 
-ByteArrayInputStream::~ByteArrayInputStream() {
+ByteArrayInputStream::~ByteArrayInputStream()
+{
 }
 
 
-void ByteArrayInputStream::close() {
+void ByteArrayInputStream::close()
+{
 }
 
 
-int ByteArrayInputStream::read(ByteBuffer& dst) {
-    if (pos >= buf.size()) {
-        return -1;
-    } else {
-        size_t bytesCopied = min(dst.remaining(), buf.size() - pos);
-        memcpy(dst.current(), &buf[pos], bytesCopied);
-        pos += bytesCopied;
-        dst.position(dst.position() + bytesCopied);
-        return bytesCopied;
-    }
+int ByteArrayInputStream::read(ByteBuffer& dst)
+{
+	if (pos >= buf.size())
+	{
+		return -1;
+	}
+	else
+	{
+		size_t bytesCopied = min(dst.remaining(), buf.size() - pos);
+		memcpy(dst.current(), &buf[pos], bytesCopied);
+		pos += bytesCopied;
+		dst.position(dst.position() + bytesCopied);
+		return (int)bytesCopied;
+	}
 }
